+++ packages/workspace-react/src/lib/components/tabNavigation/TabNavigation.tsx
@@ -1,4 +1,4 @@
-import { useTabContext } from '../Workspace';
+import { useTabContext } from '../../hooks';
 import { TabButton, TabButtonList } from './tabNavigation.styles';
 
 /**
