+++ packages/workspace-react/src/lib/hooks/useTab.ts
@@ -1,4 +1,14 @@
-import { useTabContext } from '../components/Workspace';
+import { TabController, TabProvider } from 'lib/components/Workspace';
+import { useContext } from 'react';
+import { useStore } from 'zustand';
+
+export const useTabContext = <U>(selector: (s: TabController) => U): U => {
+	const storeApi = useContext(TabProvider);
+	if (!storeApi) {
+		throw new Error('TabContext out of bounds');
+	}
+	return useStore(storeApi, selector);
+};
 
 export const useTabs = () => useTabContext((s) => s.tabs);
 export const useSetActiveTab = () => useTabContext((s) => s.setActiveTab);
